Finish up windowing (to some extent), and add it into the demo

+++ engine/include/engine/graphics/eventReturnTypes.hpp
@@ -5,8 +5,8 @@ namespace pheonix {
 
         struct XyData
         {
-            double x;
-            double y;
+            int x;
+            int y;
         };
 
     }
